Added liftLockOrConfirmDiscardingTextArea
Trying to automatically lift wiki locks on leaving an unchanged wiki edit view.

This is not working for me yet because the javascript tags in the edit_area don't seem to get triggered. The wiki toolbar is gone as well.

The ajax request I added here does not have a controller counterpart yet anyway. This is just for testing purposes.