moved js calls into edit.rjs, not using update => wiki_area anymore
As the js was not getting triggered when inserting it into the wiki_area i moved to .rjs style ajax responses. Now at the wiki toolbar is rendering and the confirmation dialogue pops up. Turns out it's always asking me to confirm even if the wiki has not been changed.