unlocking on leaving a page is working now
Using an synchronous ajax call here as the async one would sometimes be received by the controller and sometimes not. Most likely the page was unloaded faster than the request send. Not sure this is a good idea though as it might keep people from closing their browser until the server responded. Pretty rude behaviour. But then again they get a script error after a timeout. Another option might be to sleep for 0.1 seconds or so. But this feels so arbitrary.#