got tabs to work - wiki_action_links always outside div[@wiki] in group wikis
for the tabs to work two things were needed:
 * show_tab option is usually used to display static content - now
   we use it because it also toggles the tabs
 * the ajax workflow for the ajax actions needed to be cleaned up

I also looked into hiding the sidebar for the versions display for a page wiki. However this is not easy to do in terms of js / css magic and I'm also not sure if we want it.

Some css tweaks will still be required for the version display to look good.

Also Diffs still use full width normal requests. I would like to make this ajax as well. Maybe even add a View Changes tab on top instead of all these links in the versions list.

In that case i think the View Changes and Versions tab should always relate to the current diff or version shown. Might require a seperate js function.