fixed the renaming issue - needs refactoring
So this is working for now but it's terrible how we wrapped the wiki_structure in a thousand wrappers and only access it through @wikis api. I think this needs some serious refactoring. Instead of
@section = @wiki.update_section!(@section, other_args*)
I would like to do
@section.update!(other_args*)

We should have a real sections class - probably just green_tree and extend it a bit and then deal with section objects rather than just strings.