we need to rerender the whole wiki on an update
The numbers of sections with the same title can change even outside of what we are editing.
Imagine we have a wiki like this:
h1. project A
h1. project B
h2. current state

now imagine we add a 'current state' heading to the first section. The links of the h2 in the second section would have to change as we now reference this section as current_state_2.

Any other way to achieve this feels terribly cumbersome.