replaced options to the permission call with guard_like
guard_like will set the object and the verb for permission calls to use by default.

using options for permissions for this was not ideal because the permission inclusions will be inherited by subclasses whereas the class variable permission_options will not.

This way the two can be set independently. Also I added a permission_helper method to include permissions for reference in the templates but not in the controllers.