make Gallery.add_attachment! behave like add_image!
This way we can reuse the page_attachement_controller for adding images to galleries.

Also added tests that specify the new behaviour.
One test is still failing. We need to decide what we want to do when someone tries to add an image to a different gallery. Duplicates are not allowed so we could
a) move it
b) raise PermissionDenied