dealing with missalignment issues when hovering over a tab
When hovering over an active tab the bootstraps border settings would cause the line height to grow by a pixel messing up the layout.

fixed this by making the 0px border below the link !important.

Added 'reloadable' class to tab links to mark links that can be clicked when active