fixed a bunch of smaller group wiki issues - fixes #3897
When viewing group wiki publicly (not logged in) it should keep the same format as when you are logged in. I think it should say "group wiki" and then only show the public pill button.
when i edit in public wiki and hit save, i should land in the show tab of the public wiki
more and less links working now