councils keys should not be included when testing permissions for a group
castles respond to holders. users respond to holder_codes. The two are very different. holder_codes has all the codes for holders the user has access to. castle#holders are all the holders that have access to a given castle. We should not include the latter when a castle is used like a user (testing permissions it has on other castles or itself). Otherwise it will inherit permissions from holders with access to itself.