fix stylesheet and javascript inheritance in controllers
We were inheriting the hash in production and altering it in place. No idea why this happened in production but not in development.

We now use class_attributes and a setter method with the old syntax. No need for a getter anymore as the class_attributes can be read directly.