hide the owner on user and group landing pages
It's the same one anyway as these lists are filtered by owner.

We do this via css so we can still use the cached page columns.
Each row has a class corresponding to its content. (.owner in this case)
Wrapping a table in a .without_owner tag will hide those columns.
It affects both the th as well as the td so they still line up nicely.