Make sure that user_participations of page creators always have a changed_at attribute set.
Pages for which the corresponding user_participation of the current user doesn't have
a changed_at attribute set, don't show up on that user's "Recent Pages" section on
the dashboard.
I'm not sure if that is intended, but it seems so. Now newly created pages show up
again.
Also I removed User#may_admin?(thing), as it wasn't called anywhere but in Page#build!,
where it didn't work correctly.