combine page and user post permissions
We had two conflicting may_create_post? and almost conflicting may_update_post? functions. They lived in the me and the page permissions.

We often have to load both to render the menus and all actions properly. Instead of having the two fight we now have posts_permission. It's only required in a few controllers.