fix dispatch controller test, add user not found test
I want to ensure the response to a non existing context is the same as the
one to a hidden one. However this is not easy when functional testing the
dispatch controller. ActionController::TestCase will call process directly
instead of dispatch. So no way of testing the specific behaviour of
the dispatch controller.

Instead i now test the people/home controller with a hidden user and a
wrong username.

Integration and unit testing probably is the way to go for the dispatch controller.