include all styles and js and cache header
We don't have any as_needed styles or scripts that would be worth excluding.

Caching the head saves ~25 ms on each request. Not that much but still more than the 5 ms the data transfer of the entire css file takes. Further more these will be cached on the client anyway.