deprecated: class_inheritable_accessor -> class_attribute
The two behave differently but we were not using the inheritance of thumbnail definitions anyway.

So I added a test that ensures we can inherit thumbdefs without pushing the into the super class.

Documentation about the difference between the old class_inheritable_* and the new class_attribute can be found here:
http://martinciu.com/2011/07/difference-between-class_inheritable_attribute-and-class_attribute.html