remove defaults from page_creation route
It looks like routing will remove all parts that match the defaults from the route and thus make it impossible to specify :type if :action or :owner match the defaults in
  '/pages(/:action(/:owner(/:type)))'

There's no need for these defaults anyway.

Also remove the :owner => params[:owner] in new_page_path as that will
be added by default.