fix non deterministic issue with wiki integration test
So what was going on here...

we were clicking on the edit link, filling in the form and then clicking on submit.
But sometimes the page had not loaded the edit form after we clicked on the link so we would fill the textarea already loaded. Then the new textarea would be loaded and we would click submit for an empty form.

lesson learned: always wait for what you clicked on.

Actually we should prevent the same thing in the ui. But that is for later.