use holder.all_codes to aggregate default access
If there are no keys present we need to calculate the default access.

So far this was based on associated holders stored in the definition.
This allowed for customizing the default access by overwriting a method etc.

However it's fundamentally different from how we get access via keys. We already
know the set of all key_holders that might be used to get access to a castle.
So we can check if they have access by default.