test for missing piece in castle gates puzzle
Once there is a key we do not use any default access anymore.
This is okay for the holder of the key itself. However it should not
apply to other association holders that do not have a key stored yet...

In order to make this understandable... here's an example:

* user has explicit access to castle via a key that does not include :tunnel
* user is part of a group that by default has :tunnel access to the castle.

Since the user has some key only that key will be checked and we won't progress
to check the defaults of other key holders the user could use.