fix page creation test
users need to be visible to show up in autocomplete.
But you can still pester other users. So in order to test that
we may not rely on autocomplete entirely.

Also elements will loose focus as soon as they have been filled by
capybara. That way the autocomplete will not show up. So we need
to make sure they stay focussed.