style and clean up people directory autocomplete
We now use a form with data-remote: true and submit that form via ajax
when an element is selected from the autocomplete. This makes the whole
autocomplete a lot less complex. No more need to trigger remote_functions
from handlers.

I hope to extend this approach to all autocompletes. It will also help us
when we migrate to jquery as it relies a lot less on prototype.