replace autocomplete_users_field_tag
use autocomplete_recipients_field_tag in message form
use autocomplete_input_tag in other places.

The former will only list users that may be pestered.
The latter will autosubmit the form if an entry is selected.

Autosubmit is fine for the search_filters we're using this for here. However we don't really submit() the form. Instead we trigger a click on the submit button so it can trigger a remote function. We will replace this with a less obstrusive js version later.