deal with complex headings in a meaningful way - fixes #4151
The regexp for detecting the headings has become a lot more loose.
To compensate for that we now start the scan at the parent headings line.

In theory it should be enough to just find the next heading. We still make sure it includes all words though.

We also won't catch errors in greencloth anymore. Instead they will escelate into the app itself and it can log them and deal with them.