upgrade: tweak MessagePageConversion
* turn MessagePages with assets into DiscussionPages right away.
* print dots every 10 MessagePages as the whole process takes loong.
* dismiss all the new PrivateMessageNotices

This is not ideal as the process is still slow and some of the
messages might have been unread and still the notice is dismissed.

However - the whole thing is slow because we are replacing one complex
data structure (page with all the records attached to it) with another
complex data structure (discussion + post + notice). So we actually
instantiate all the records and destroy / create all the things.

I don't see a good option to keep the unread state of the messages either.