performance: tweaks for user permission migration
Instead of fetching the castle for a key and then looking at its class for the gate_set we now constantize the castle_type.

When the keys of a castle are loaded we now detect the one with the appropriate holder_code instead of fetching it from the database.

We also allow for preloading public and private_profile