encode + in group names in share autocomplete
+ is a way to encode ' '. So it will be replaced by rails automagically.
So we have to encode it when sending the first request.
In the form it will be encoded anyway. So no need to replace it in the hidden input.

While we are at it... also display the display_name on top and in bold and the name below

I may have found a way to fix the annoying non determenistic autocomplete integration
test failure at the same time. One thing is for sure ' r' in the field will trigger
the first request to autocomplete - even if the thing we are looking for is already
seeded. So instead i now try to fill in the space first - to prevent the first char
from getting cut off and then fill in only the term to autocomplete.