use custom subclass of AD::PublicExceptions
For a 404 it will use our own controller and fall back to the standard pages in public.
Everything else still uses the ActionDispatch::PublicExceptions. We still try to rescue
everything else within the app anyway.
In case there is a 500 (or 422) we do not rely on the ExceptionsController but use a plain
file for now.