performance: drop posts_count from pages
It's unused right now and if we will display posts counts again caching the
corresponding partial is probably a better idea.

Also it was causing pages to get saved twice when creating a post. Obviously this
could have been solved differently - however since we do not use this right now i
just removed it.