use has_secure_password and wrap old sha1 hashes
has_secure_password uses bcrypt which is far better at hashing passwords
than sha1.

In order to profit from the security gains of bcrypt without waiting for
all users to login again so we can encrypt their plaintext password we use
bcrypt(sha1(password, salt)) as an intermediate step. This allows us to
simply bcrypt all existing hashes. We obviously need to keep the old salt.

When a user logs in we calculate the old hash like before and then use the
password comparison of has_secure_password.
If the password worked we also rehash it with plain bcrypt and remove the old
salt from the database. This way we can tell how a password was hashed by
looking at the salt of the record.

Once all users have logged in or their accounts have expired we can remove
these hacks.

ATTENTION:
 * migration required
 * rake task cg:upgrade:secure_password required