test: fix tests that failed 1 hour before midnight
We process trackings of user action step by step. First we add those from within an hour to sum them up to an 'Hourly'. Then after 24 hours we sum up the hourlies to a 'Daily'.

Now in order to test records with Dailies the corresponding tracked action needs to be at least from the day before yesterday so it will be compressed into dailies. For this we used Time.now - 2.days. However the hourlies are stored with created_at set to the end of the hour. So Time.now - 2.days and then the end of the hour actually is midnight - so the start of yesterday - which won't be counted in the dailies. So instead we now use Time.now - 3.days to make sure we actually produce some proper dailies.