bring back ErrorNotFound, drop per site i18n
ErrorNotFound is similar to ActiveRecord::RecordNotFound - but it allows
specifying the type of thing that was not found for more detailed error
messages.

For the translations we use cascading translations. So say a group was not
found... we will lookup the following:
exception.title.group.not_found
exception.title.not_found
exception.not_found
not_found

For all of these %{thing} will be interpolated with a translation of :group.
If there is no translation for the given class of things it will be an empty
string.

In order for the cascading translation to work we have do drop the support
for site specific translations. Otherwise the lookup will start with a
cascade from:
site.exception.title.group.not_found
and succeed when looking up 'not_found'.
So it won't process the more detailed cascade correctly.

Alternatively we could add the site scope last. But I think site specific
translations are not something we will need any time soon and it adds a lot of
complexity. Instance specific translations can easily be put in
  /extensions/locales
and overwrite the defaults.