Notification class to notify in different ways
A notification informs users about an action that just took place.

The users may be notified by different means - email, notice, etc.
For now this will be hardcoded but it may be configurable for users at
some point.

For each thing that happens one notification will be instantiated. It will
then inform all the recipients and may create many emails or notices.