refactor: Add users to group they created in controller
Observers are deprecated and User.current is a bad practice.
Instead let's be explicit about the steps involved in the controller.

Also ensure users do not get added to networks by default

Networks are created with an initial group member. If the user gets access to the network via this group there is no need to add them directly.