prevent 500 for links that still use url encodes iso chars
We used some iso encoding before and that would turn some charectars into
%F3 and similar when url encoded.

Now that we switched to utf-8 some of those links still exist but lead to
utf-8 errors as characters above \xF0 are invalid. Since the exact filename
is not relevant for the assets controller anyway and this has been where we
saw the issue in production we will just drop the invalid chars in the utf-8
encoding.