remove user observer, prepare user destroyed notices
Preparing to user UserDestroyedNotice rather than UserDestroyedActivity.
Activities should have a subject the activity is about. After the user
was destroyed it can't be about that particular user anymore. So instead
the subject was the recipient. Very confusing. We now use Notices for
notifying a particular user about an event.

Notification will try to guess the Notice name for now and if it does
not exist it will log a warning.