separate unit tests - share page with one or many
I'm about to create a new model for sharing with multiple users. As a
first step let's separate the tests.

What's the plan?

I want to create the PageHistory::GrantAccess notifications in the
Pages::SharesController. Therefore I need the set of pariticipations
that were created or updated. While this would be feasible within
the current UserExtension::Pages our User model already almost does
everything so let's take this out of the User model into a into a
small utility class with immutable instances that trigger the individual
sharing and collect the participations.