fix participation tracking, actually use it
Turns out we were still using the old GroupParticipationObserver.
Not that is gone.

We also set the item and access via participation setters now. That's
probably cleaner and works with filtering the attributes in Tracking::Action.

Oh... and the participations are saved BEFORE we run the track_action after
filter - so we have to look at previous_changes rather than access_changed.