make comments and messages more robust
We now use two controller actions ... one for generating the post and
then we redirect to one for rerendering the list of posts.

This way we can set @post to the created post in the creation action
and track it and have a new Post as @post in the index action.