Exception: Translation missed for network group invite notification
When I invite a group A to join the network B, the members from this group A,
see an exception on `/me` page.

The exception description notices:
"I18n::MissingTranslationData: translation missing: en.request_to_join_our_network"

It occurs because of the notification constructed for user with info about request to join network.

This issue related to the fixed issue "There is no notification for invite to join group" [https://labs.riseup.net/code/issues/9676], because before we have no such notifications at all..

So need to be tested only after #9676 would be resolved and merged.

Bug: 9686
Link: https://labs.riseup.net/code/issues/9686