enable wiki diffs for group wikis
Interestingly the diff will also show if you changed the wiki yourself.
Why?
Because the wiki update happens in a different controller that does not
set visited_at to a new date. In contrast to the page wiki where updates
always mark the page as seen.

I'm not sure yet wether this is a bug or a feature. Maybe we should always
highlight the diff after one saved a wiki?