Anchor links for comments
Here is fix for one issue, related to comment date link
and small feature implementation.

Issue: if we edit a comment that we own its date time becomes a link
that supposed show real time created/modification on click.
This functionality doesn't work at all, leads to javascript error in browser console.

Feature:
Now if I want to point somebody on some specific comment I could only note
the page where it is posted, but after user should go though all of the comments
to find an appropriate.

It is not comfortable and decreases quoting possibilities if we want
share links to comments on external resourses.

So I think that it might be useful.

Issue: #9729
Link: https://labs.riseup.net/code/issues/9729