bugfix: duplicate use of section names broke wiki
A wiki with the same section names nested inside each other broke
redcloth because the heading regexp was to loose.

We do a depth first search for headings. Looking for the nested heading would then pick the parent heading with the old regexp. In particular we always searched for underlined headings even if the heading level was higher than 2 and no underlines were possible. We also searched for underlines with '==' and '--' no matter if the heading was an h1 or and h2.

This is now fixed so that only the proper underlines will match.