bugfix: 500 on trying to convert xcf
We currently do not have a transmogrifier for xcf. This fixes two different
issues:

Thumbnail records were created despite not having the transmogrifier.
I think this was caused by Media::Transmogrifier returning an IO
object because that's what log_error returns.

Now we have these dangling Thumbnail records in the database and have to
deal with the issue they cause.

When loading the asset we try to create the thumbnails that do not exist
yet. While this is fine for thumbnails we have a transmogrifier for it fails
otherwise. This was causing the requests for the files to err out with a 500.