try converting first page of pdf only
It's way faster for large documents. If it fails we retry without setting the page.

You can use filename.pdf[0] in gm convert to only convert the first
page. This speeds up dealing with large documents a lot.
(.5 sec instead of 18 sec. for a 80 page, 7MB pdf i tried).

Rumors have it that some documents use letters for indexing their pages
and [0] won't work. So we need a fallback.