refactor: separate crabgrass_media repository
crabgrass_media depends on a number of packages to be installed.
We don't install them on travis right now because that makes test
runs much faster.

But we still need to test crabgrass_media with these packages properly.
So this will now live in a different repo and be tested on its own.