revert wiki decoration changes, move to class
I tried to fix a bunch of issues that came up during wiki decoration lately.
Turns out i completely broke section editing in the end.

So to prevent this for the future we now have an integration test.

I also moved the bulk of the wiki decoration from the helper to a class of
its own - Wiki::Decorator. This way we can unit test the different issues
that come up instead of relying on integration tests.