performance: cache wiki template
The wiki/wiki/_show template is by far the one that takes most of the total
time spend rendering views.

This comes as no surprise as we not only insert the wiki.body_html but parse it
and extend it with the edit links and so on.

So instead of doing this all the time we now cache it based on the wiki (updated_at),
and the users locale if they are allowed to edit the wiki. The latter is necessary
for translating the wiki edit links and their titles.