fix umlaut issues in page titles and anchors
Moving from ruby 1.8 to 1.9 \w does not include non ascii chars anymore.
Now we need to use [[:word:]] to account for utf-8 chars in the String#nameize.

This came up in wiki section anchors but also affected page titles.
I added an integration test for the page title.

The test failed due to invalid url errors when capybara handed the url to
URI.parse. So now we URI.encode the page_path. While this does change the
string in the test iceweasel still displays the umlaut characters nicely.