use thinking sphinxs delayed delta indexing
Thinking Sphinx will now create a delayed job for every indexing task.

This way the indexing will happen in a separate worker task and rails
can respond faster.

In order for this to work the delayed job tasks need to be run. That's
what the script/delayed_job script is for. You can start it as a daemon:
```
  RAILS_ENV=production script/delayed_job start
```

Running the delta indexing asynchronous also reduces the need for
reindexing. The delta index may grow and delta indexing may be slow -
but it won't delay the response anymore.

Therefore this commit also sets the reindex frequency to 6 hours. Here's
the math behind this:

R denotes the time the reindex takes and d denotes the additional time
required by a doc in the delta index for a single index run. Then the
average index time per document when growing the delta index up to n
documents is:

 t(n) = R/n + d(n+1)/2

This function has a minimum at n_opt = sqr(2R/d). For our production install
R = 80s and d = 2.5 ms. So n_opt ~ 180 documents. We should have a few
less in 6 hours. So there is room to grow.

I hope to speed up reindexing with a new version of sphinx and higher
mem_limit. Then recalculating this might be interesting. But for now
sphinx should not be a performance bottle neck anymore.