bugfix: make reindex cron action work
The Rails.root path would have required a .to_s. But the whole
idea of running the rake task via system makes no sense in the
context of the CronController.

Instead we now use whenevers rake type.

rake ts:index takes 80 seconds total, 3 of those are for startup.
No point in blocking a passenger process for 80 seconds to avoid
the 3 second startup.

this superseeds pr #307 on github