bugfix: make reindex cron action work
The Rails.root path would have required a .to_s. But the whole
idea of running the rake task via system makes no sense in the
context of the CronController.

As the comment above says:

So there is no point in using this mechanism to run system calls.

I'm wondering though if it's a clever idea to require rake and
the thinking sphinx tasks in a controller.