tasks: fix unassigning last user
if all checkboxes are unchecked the browser does not submit the param
at all instead of submitting an empty array.

So if the param is not present we have to assume it's an empty array.
This is odd and may break intuition if anyone wants to talk to the
server directly. But i don't see any better way to fix this right now.