cron: drop cache cleanup tasks
There's no point in triggering a request with curl to then call
system from the controller.

Also... we want to keep fragment caches for more than just 3 days.
Currently we have about a month of caches and they occupy a total of
< 800MB. Many of them are probably page list entries and i bet they still are
used even if they have not been updated in a while.