start to bring back notification emails
We did not send them for a while and now we have a huge backlog.
So processing the backlog with the old routines has become impossible.

We used to load all page_histories into memory that had no notification send
and then process them. Currently these are > 1,000,000 records. So it eats
all mem and overwhelms the server.

Instead we now send the single notification using delayed job as pages are changed.
Digest notifications are not working for now. We might want to redesign them
anyway.