tests: some versioning fixes
One test was failing because the version setup did not work anymore. Instead of 6 versions it would create 5 as it tried to create two versions as the same user.

I also rewrote a different test which i used to track the problem down.

I also rewrote Wiki#create_new_version? a bit. It still does the same.
I found the old code hard to grasp - i hope this is a bit more clear.