prevent duplicate asset upload
assigning a name to the upload-input handler will also submit the file
as a parameter for the given key. We add the file to the upload as the
asset[uploaded_file] with javascript so we end up uploading it for two
destinct keys and thus uploading it twice. :open_mouth: