Use InnoDB for all tables but page_terms
InnoDB allows us to perform non blocking backups with transactions.

So we convert all the tables that remained as MyISAM tables to
InnoDB except page_terms which need the full text search capacities
that InnoDB only gained with mysql 5.6. (we're still on 5.5).

trackings were actively using "INSERT DELAYED" which also is
MyISAM specific - but will be removed in this commit as well.
The mysql docs say:

> Note that INSERT DELAYED is slower than a normal INSERT if the table
> is not otherwise in use.

So it very much looks like premature optimization.

The other tables just happen to still be MyISAM in production